opgeruimd + exerciseDescriptions worden nu vanuit json gelezen
