+++ src/exerciseDataCreator/BodyPartDescriptionType.py
