+++ src/exerciseDataCreator/ImpairedSideType.py
