+++ src/exerciseDataCreator/PoseDetectionType.py
