+++ src/exerciseScorer/PoseDetectionType.py
