+++ src/poseDetection/BodyPartType.py
