+++ src/poseDetection/HandPartType.py
